// front_/src/app/models/document.enum.ts
export enum TypeDocument {
    ARTICLE = 'ARTICLE',
    VIDEO = 'VIDEO',
    PHOTO = 'PHOTO',
    PRESENTATION = 'PRESENTATION',
    REPORT = 'REPORT',
    RESEARCH_PAPER = 'RESEARCH_PAPER',
    EBOOK = 'EBOOK',
    TUTORIAL = 'TUTORIAL',
    THESIS = 'THESIS'
}

export enum TypeDocumentVisibility {
    PUBLIC = 'PUBLIC',
    PRIVATE = 'PRIVATE'
}

export enum TypeAccessLevelDocument {
    READ = 'READ',
    WRITE = 'WRITE'
}

export enum TypeDocumentTheme {
    TECHNOLOGY = 'TECHNOLOGY',
    MEDICINE = 'MEDICINE',
    HEALTHCARE = 'HEALTHCARE',
    ENTREPRENEURSHIP = 'ENTREPRENEURSHIP',
    BUSINESS = 'BUSINESS',
    ENGINEERING = 'ENGINEERING',
    SCIENCES = 'SCIENCES',
    ARTS = 'ARTS',
    DESIGN = 'DESIGN',
    EDUCATION = 'EDUCATION'
   
}


  